refactor(backoffice-v2): moved aml cells to a re-usable hook
